Move scanFace components into folder
