+++ src/components/scanFace/RegistrationPage.tsx
